% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_re_exports.R
\name{base functions}
\alias{base functions}
\title{Pseudo-function to re-export functions from the \strong{stats} package.}
\description{
Pseudo-function to re-export functions from the \strong{stats} package.
}
