% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_london_normalised.R
\docType{data}
\name{data_london_normalised}
\alias{data_london_normalised}
\title{Example meteorologically normalised data for the \strong{rmweather} package.}
\format{Tibble with 246 observations and 5 variables. The variables are:
\code{date}, \code{date_end}, \code{site}, \code{site_name}, and 
\code{value_predict}. The dates are in \code{POSIXct} format, the site 
variables are characters and \code{value_predict} is numeric.}
\usage{
data_london_normalised
}
\description{
These example data are derived from the observational data included in 
\strong{rmweather} and represent meteorologically normalised NO2 
concentrations at London Marylebone Road, aggregated to monthly resolution.
}
\examples{

# Load rmweather's meteorologically normalised example data and check
head(data_london_normalised)

}
\seealso{
\code{\link{data_london}}
}
\author{
Stuart K. Grange
}
\keyword{datasets}
