% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_calculate_model.R
\name{met_calculate_model}
\alias{met_calculate_model}
\title{Function to}
\usage{
met_calculate_model(df, variables, trees = 300, mtry = NULL,
  min_node_size = 5, n_cores = NULL, verbose = FALSE)
}
\description{
Function to
}
