% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_normalise.R
\name{met_normalise}
\alias{met_normalise}
\title{Function to normalise a variable for "average" meteorological conditions.}
\usage{
met_normalise(model, df, variables = NA, n_samples = 300, replace = TRUE,
  n_cores = NA, verbose = FALSE)
}
\arguments{
\item{model}{A ranger model object from \code{\link{met_train_model}}.}

\item{df}{Input data used to calculate \code{model} using 
\code{\link{met_prepare_data}}.}

\item{variables}{Variables to randomly sample. Default is all variables used
for training the model with the exception of \code{date_unix}, the trend term.}

\item{n_samples}{Number of times to sample \code{df} and then predict?}

\item{replace}{Should \code{variables} be sampled with replacement?}

\item{n_cores}{Number of CPU cores to use for the model predictions. Default
is system's total minus one.}

\item{verbose}{Should the function give messages?}
}
\value{
Data frame.
}
\description{
Function to normalise a variable for "average" meteorological conditions.
}
\examples{
\dontrun{

# Normalise a time series
data_normalised <- met_normalise(
  model, 
  data_for_modelling, 
  n_samples = 300,
  verbose = TRUE
)

}

}
\seealso{
\code{\link{met_prepare_data}}, \code{\link{met_train_model}}
}
\author{
Stuart K. Grange
}
