% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_plot_partial_dependencies.R
\name{met_plot_partial_dependencies}
\alias{met_plot_partial_dependencies}
\title{Function to plot partial dependencies after calculation by 
\code{\link{met_partial_dependencies}}.}
\usage{
met_plot_partial_dependencies(df)
}
\arguments{
\item{df}{Data frame created by \code{\link{met_partial_dependencies}}.}
}
\value{
ggplot2 plot.
}
\description{
Function to plot partial dependencies after calculation by 
\code{\link{met_partial_dependencies}}.
}
\author{
Stuart K. Grange
}
