% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_predict.R
\name{met_predict}
\alias{met_predict}
\title{Function to predict.}
\usage{
met_predict(model, df = NA, n_cores = NULL, verbose = FALSE)
}
\arguments{
\item{model}{A ranger model object from \code{met_train_model}.}

\item{df}{Input data to be used for predictions.}

\item{n_cores}{Number of CPU cores to use for the model predictions.}

\item{verbose}{Should the function give messages?}
}
\value{
Numeric vector.
}
\description{
Function to predict.
}
\author{
Stuart K. Grange
}
