% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_predict_the_test_set.R
\name{met_predict_the_test_set}
\alias{met_predict_the_test_set}
\alias{met_plot_test_prediction}
\title{Functions to use a model to predict the observations within a test set after
\code{met_calculate_model}.}
\usage{
met_predict_the_test_set(model, df)

met_plot_test_prediction(df)
}
\arguments{
\item{model}{A ranger model object from \code{met_calculate_model}.}

\item{df}{Input data used to calculate \code{model}.}
}
\value{
Data frame.
}
\description{
\code{met_predict_the_test_set} uses data withheld from the training of the 
model and therefore can be used for investigating overfitting.
}
\examples{

\dontrun{

# Use the test set for prediction
met_predict_the_test_set(model, data_for_modelling)

# Produce a hex plot of the predictions
met_predict_the_test_set(model, data_for_modelling) \%>\% 
  met_plot_test_prediction()
 
}

}
\author{
Stuart K. Grange
}
