% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_prepare_data.R
\name{met_prepare_data}
\alias{met_prepare_data}
\title{Function to prepare a data frame for modelling with \strong{metnormr}.}
\usage{
met_prepare_data(df, value)
}
\value{
Data frame, the input data transformed to a data frame ready for 
modelling with \strong{metnormr}.
}
\description{
\code{met_prepare_data} will check if a \code{date} variable is present and 
is of the correct data type, impute missing numeric and catagorical values, 
randomly split the input into training and testing sets, and rename the 
dependant variable to \code{"value"}.
}
\examples{
\dontrun{

# Prepare data for modelling 
data_for_modelling <- met_prepare_data(data_london, value = "no2")

}

}
\author{
Stuart K. Grange
}
