% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_london.R
\docType{data}
\name{model_london}
\alias{model_london}
\title{Example \strong{ranger} random forest model for the \strong{rmweather} 
package.}
\format{
A ranger object, a named list with 14 elements.
}
\usage{
model_london
}
\description{
This example object was created from the observational data included in 
\strong{rmweather} and is a random forest model returned by 
\code{\link{rmw_train_model}}. This forest is only made from one tree to keep
the file size small and is only used for the package's examples.
}
\examples{

# Load rmweather's ranger model example data and see what elements it contains
names(model_london)

# Print ranger object
print(model_london)

}
\seealso{
\code{\link{data_london}}, \code{\link{rmw_train_model}}
}
\author{
Stuart K. Grange
}
\keyword{datasets}
