% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_re_exports.R, R/pipe_re_export.R
\name{\%>\%}
\alias{\%>\%}
\alias{\%>\%}
\title{Psudo-function to re-export \strong{magrittr}'s pipe.}
\description{
Psudo-function to re-export \strong{magrittr}'s pipe.

Psudo-function to re-export \strong{magrittr}'s pipe.
}
