% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release_questions.R
\name{release_questions}
\alias{release_questions}
\title{Function to print messages before a CRAN submission.}
\usage{
release_questions()
}
\description{
Function to print messages before a CRAN submission.
}
\details{
Not to be exported.
}
