% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmw_calculate_model_errors.R
\name{rmw_calculate_model_errors}
\alias{rmw_calculate_model_errors}
\title{Function to calculate observed-predicted error statistics.}
\usage{
rmw_calculate_model_errors(
  df,
  value_model = "value_predict",
  value_observed = "value",
  testing_only = TRUE,
  as_long = FALSE
)
}
\arguments{
\item{df}{Data frame with observed-predicted variables.}

\item{value_model}{The modelled/predicted variable in \code{"df"}.}

\item{value_observed}{The observed variable in \code{"df"}.}

\item{testing_only}{Should only the testing set be used for the calculation 
of errors?}

\item{as_long}{Should the returned tibble be in "long" format? This is useful
for plotting.}
}
\value{
Tibble.
}
\description{
Function to calculate observed-predicted error statistics.
}
\author{
Stuart K. Grange
}
