% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmw_example_data.R
\name{rmw_example_data}
\alias{rmw_example_data}
\title{Function to load example data for the \strong{rmweather} package.}
\usage{
rmw_example_data()
}
\value{
Data frame.
}
\description{
Function to load example data for the \strong{rmweather} package.
}
\author{
Stuart K. Grange
}
