% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmweather_plotting_functions.R
\name{rmw_plot_importance}
\alias{rmw_plot_importance}
\title{Function to plot random forest imporantances after training by 
\code{\link{rmw_train_model}}.}
\usage{
rmw_plot_importance(df)
}
\arguments{
\item{df}{Data frame created by \code{\link{rmw_model_importance}}.}
}
\value{
ggplot2 plot.
}
\description{
Function to plot random forest imporantances after training by 
\code{\link{rmw_train_model}}.
}
\author{
Stuart K. Grange
}
