% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmweather_plotting_functions.R
\name{rmw_plot_normalised}
\alias{rmw_plot_normalised}
\title{Function to plot the meteorologically normalised time series after
\code{\link{rmw_normalise}}.}
\usage{
rmw_plot_normalised(df, colour = "#6B186EFF")
}
\arguments{
\item{df}{Data frame created by \code{\link{rmw_normalise}}.}

\item{colour}{Colour for line geometry.}
}
\value{
ggplot2 plot with a line geometry.
}
\description{
Function to plot the meteorologically normalised time series after
\code{\link{rmw_normalise}}.
}
\examples{

# Plot normalised example data
rmw_plot_normalised(data_london_normalised)

}
\author{
Stuart K. Grange
}
