% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmweather_plotting_functions.R
\name{rmw_plot_partial_dependencies}
\alias{rmw_plot_partial_dependencies}
\title{Function to plot partial dependencies after calculation by 
\code{\link{rmw_partial_dependencies}}.}
\usage{
rmw_plot_partial_dependencies(df)
}
\arguments{
\item{df}{Data frame created by \code{\link{rmw_partial_dependencies}}.}
}
\value{
ggplot2 plot with a point geometry.
}
\description{
Function to plot partial dependencies after calculation by 
\code{\link{rmw_partial_dependencies}}.
}
\author{
Stuart K. Grange
}
