% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmweather_plotting_functions.R
\name{rmw_plot_test_prediction}
\alias{rmw_plot_test_prediction}
\title{Function to plot the test set and predicted set after
\code{\link{rmw_predict_the_test_set}}.}
\usage{
rmw_plot_test_prediction(df)
}
\arguments{
\item{df}{Data frame created by \code{\link{rmw_predict_the_test_set}}.}
}
\value{
ggplot2 plot.
}
\description{
Function to plot the test set and predicted set after
\code{\link{rmw_predict_the_test_set}}.
}
\author{
Stuart K. Grange
}
