% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmw_predict.R
\name{rmw_predict}
\alias{rmw_predict}
\title{Function to predict using a \strong{ranger} random forest.}
\usage{
rmw_predict(model, df = NA, n_cores = NULL, verbose = FALSE)
}
\arguments{
\item{model}{A \strong{ranger} model object from \code{rmw_train_model}.}

\item{df}{Input data to be used for predictions.}

\item{n_cores}{Number of CPU cores to use for the model predictions.}

\item{verbose}{Should the function give messages?}
}
\value{
Numeric vector.
}
\description{
Function to predict using a \strong{ranger} random forest.
}
\examples{
\donttest{

# Make a prediction with a ranger random forest model
vector_predict <- rmw_prefict(model, df = data_observations)

}

}
\author{
Stuart K. Grange
}
