% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmweather_helpers.R
\name{system_cpu_core_count}
\alias{system_cpu_core_count}
\title{Function to return the system's number of CPU cores.}
\usage{
system_cpu_core_count(logical_cores = TRUE, max_cores = NA)
}
\arguments{
\item{logical_cores}{Should logical cores be included in the core count?}

\item{max_cores}{Should the return have a maximum value? This can be useful 
when there are very many cores and logic is being built.}
}
\description{
Function to return the system's number of CPU cores.
}
\author{
Stuart K. Grange
}
