% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metnormr_helpers.R
\name{wday_monday}
\alias{wday_monday}
\title{Function to get weekday number from a date where \code{1} is Monday and 
\code{7} is Sunday.}
\usage{
wday_monday(x, as.factor = FALSE)
}
\arguments{
\item{x}{Date vector.}

\item{as.factor}{Should the return be a factor?}
}
\value{
Numeric vector.
}
\description{
Function to get weekday number from a date where \code{1} is Monday and 
\code{7} is Sunday.
}
\author{
Stuart K. Grange
}
