% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_normalise.R
\name{met_normalise}
\alias{met_normalise}
\title{Function to normalise a variable for "average" meteorological conditions.}
\usage{
met_normalise(model, df, n_samples = 300, replace = TRUE, n_cores = NULL,
  verbose = FALSE)
}
\arguments{
\item{model}{A ranger model object from \code{met_calculate_model}.}

\item{df}{Input data used to calculate \code{model}.}

\item{n_samples}{Number of times to sample \code{df} and then predict?}

\item{replace}{Should \code{variables} be sampled with replacement?}

\item{n_cores}{Number of CPU cores to use for the model calculation.}

\item{verbose}{Should the function give messages?}

\item{variables}{Variables to randomly sample. \code{variables} will be all
variables used for calculating the model in \code{met_calculate_model} with
the exception of \code{date_unix}, the trend term.}
}
\value{
Data frame.
}
\description{
Function to normalise a variable for "average" meteorological conditions.
}
\examples{
\dontrun{

# Normalise a time series
data_normalised <- met_normalise(
  model, 
  data_for_modelling, 
  n_samples = 300,
  n_cores = 7, 
  verbose = TRUE
)

}

}
\seealso{
\code{\link{met_prepare_data}}, \code{\link{met_calculate_model}}
}
\author{
Stuart K. Grange
}
