% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_partial_dependencies.R
\name{met_partial_dependencies}
\alias{met_partial_dependencies}
\title{Function to}
\usage{
met_partial_dependencies(model, df, variable = NA, two_way = FALSE,
  n_cores = NA, verbose = FALSE)
}
\arguments{
\item{df}{Input data frame.}
}
\value{
Data frame.
}
\description{
Function to
}
\author{
Stuart K. Grange
}
