% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_train_model.R
\name{met_train_model}
\alias{met_train_model}
\title{Function to train a random forest model to predict (usually) pollutant
concentrations using meteorological and time variables.}
\usage{
met_train_model(df, variables, n_trees = 300, mtry = NULL,
  min_node_size = 5, n_cores = NA, verbose = FALSE)
}
\arguments{
\item{df}{Input data frame after preparation with 
\code{\link{met_prepare_data}}. \code{df} has a number of constraints which 
will be checked for before modelling.}

\item{variables}{Independent/explanatory variables used to predict 
\code{"value"}.}

\item{n_trees}{Number of trees to grow to make up the forest.}

\item{mtry}{Number of variables to possibly split at in each node. Default is 
the (rounded down) square root of the number variables.}

\item{min_node_size}{Minimal node size.}

\item{n_cores}{Number of CPU cores to use for the model calculation. Default
is system's total minus one.}

\item{verbose}{Should the function give messages?}
}
\value{
A \code{ranger} model object, a named list.
}
\description{
Function to train a random forest model to predict (usually) pollutant
concentrations using meteorological and time variables.
}
\examples{

\dontrun{

# Calculate a model using common meteorological and time variables
model <- met_train_model(
  data_for_modelling,
  variables = c(
    "ws", "wd", "air_temp", "rh", "date_unix", "day_julian", "weekday", "hour"
  ),
  n_trees = 300
)

}

}
\seealso{
\code{\link{met_prepare_data}}, \code{\link{met_normalise}}
}
\author{
Stuart K. Grange
}
