% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmw_example_data.R
\name{rmw_example_data}
\alias{rmw_example_data}
\title{Function to load example data for the \strong{rmweather} package.}
\usage{
rmw_example_data()
}
\value{
Data frame.
}
\description{
Function to load example data for the \strong{rmweather} package.
}
\details{
These example data are daily means of NO2 and NOx observations at 
London Marylebone Road. The accompanying surface meteorological data are from
London Heathrow, a major airport located 23 km west of Central London.
}
\examples{

# Load rmweather's example data
data_london <- rmw_example_data()


}
\author{
Stuart K. Grange
}
