% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmw_model_statistics.R
\name{rmw_model_statistics}
\alias{rmw_model_statistics}
\alias{rmw_model_importance}
\title{Functions to extract model statistics from a model calculated with
\code{rmw_calculate_model}.}
\usage{
rmw_model_statistics(model)

rmw_model_importance(model)
}
\arguments{
\item{model}{A ranger model object from \code{rmw_calculate_model}.}
}
\value{
Data frame.
}
\description{
Functions to extract model statistics from a model calculated with
\code{rmw_calculate_model}.
}
\examples{

\dontrun{

# Extract statistics from a model object
rmw_model_statistics(model)

# Extract importances from a model object
rmw_model_importance(model)
 
}

}
\author{
Stuart K. Grange
}
