% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmweather_plotting_functions.R
\name{rmw_plot_importance}
\alias{rmw_plot_importance}
\title{Function to plot random forest variable importances after training by 
\code{\link{rmw_train_model}}.}
\usage{
rmw_plot_importance(df, colour = "black")
}
\arguments{
\item{df}{Data frame created by \code{\link{rmw_model_importance}}.}

\item{colour}{Colour of point and segment geometries.}
}
\value{
ggplot2 plot with point and segment geometries.
}
\description{
Function to plot random forest variable importances after training by 
\code{\link{rmw_train_model}}.
}
\seealso{
\code{\link{rmw_train_model}}, \code{\link{rmw_model_importance}}
}
\author{
Stuart K. Grange
}
