% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmweather_plotting_functions.R
\name{rmw_plot_test_prediction}
\alias{rmw_plot_test_prediction}
\title{Function to plot the test set and predicted set after
\code{\link{rmw_predict_the_test_set}}.}
\usage{
rmw_plot_test_prediction(df, bins = 30)
}
\arguments{
\item{df}{Tibble created by \code{\link{rmw_predict_the_test_set}}.}

\item{bins}{Numeric vector giving number of bins in both vertical and 
horizontal directions.}
}
\value{
ggplot2 plot with a hex geometry.
}
\description{
Function to plot the test set and predicted set after
\code{\link{rmw_predict_the_test_set}}.
}
\author{
Stuart K. Grange
}
