% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmw_predict.R
\name{rmw_predict}
\alias{rmw_predict}
\title{Function to predict using a \strong{ranger} random forest.}
\usage{
rmw_predict(model, df = NA, n_cores = NULL, verbose = FALSE)
}
\arguments{
\item{model}{A \strong{ranger} model object from \code{rmw_train_model}.}

\item{df}{Input data to be used for predictions.}

\item{n_cores}{Number of CPU cores to use for the model predictions.}

\item{verbose}{Should the function give messages?}
}
\value{
Numeric vector.
}
\description{
Function to predict using a \strong{ranger} random forest.
}
\examples{

# Make a prediction with the examples
vector_prediction <- rmw_predict(
  model_london, 
  df = rmw_prepare_data(data_london, value = "no2")
)

}
\author{
Stuart K. Grange
}
