% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metnormr_helpers.R
\name{system_cpu_core_count}
\alias{system_cpu_core_count}
\title{Function to return the system's number of CPU cores.}
\usage{
system_cpu_core_count(logical_cores = TRUE)
}
\arguments{
\item{logical_cores}{Should logical cores be included in the core count?}
}
\description{
Function to return the system's number of CPU cores.
}
\author{
Stuart K. Grange
}
